+++ ttbasic/esp32gfx.h
@@ -215,7 +215,7 @@ public:
     return m_pixels[y][x];
   }
 
-  void MoveBlock(uint16_t x_src, uint16_t y_src, uint16_t x_dst, uint16_t y_dst, uint8_t width, uint8_t height, uint8_t dir);
+  void MoveBlock(uint16_t x_src, uint16_t y_src, uint16_t x_dst, uint16_t y_dst, uint16_t width, uint16_t height, uint8_t dir);
 
   inline void setInterlace(bool interlace) {
     //m_interlace = interlace;
